--- helpers/helpers_test.go
@@ -1,62 +1,0 @@
-package helpers
-
-import (
-	"errors"
-	"fmt"
-	"os"
-	"testing"
-
-	"github.com/senzing-garage/sz-sdk-go/szerror"
-	"github.com/stretchr/testify/assert"
-)
-
-// ----------------------------------------------------------------------------
-// Interface methods - test
-// ----------------------------------------------------------------------------
-
-func TestHelpers_GetLogger(test *testing.T) {
-	x := GetLogger(1, map[int]string{}, 4)
-	assert.NotEmpty(test, x)
-}
-
-func TestHelpers_GetMessenger(test *testing.T) {
-	x := GetMessenger(1, map[int]string{}, 4)
-	assert.NotEmpty(test, x)
-}
-
-// ----------------------------------------------------------------------------
-// Test harness
-// ----------------------------------------------------------------------------
-
-func TestMain(m *testing.M) {
-	err := setup()
-	if err != nil {
-		if errors.Is(err, szerror.ErrSzUnrecoverable) {
-			fmt.Printf("\nUnrecoverable error detected. \n\n")
-		}
-		if errors.Is(err, szerror.ErrSzRetryable) {
-			fmt.Printf("\nRetryable error detected. \n\n")
-		}
-		if errors.Is(err, szerror.ErrSzBadInput) {
-			fmt.Printf("\nBad user input error detected. \n\n")
-		}
-		fmt.Print(err)
-		os.Exit(1)
-	}
-	code := m.Run()
-	err = teardown()
-	if err != nil {
-		fmt.Print(err)
-	}
-	os.Exit(code)
-}
-
-func setup() error {
-	var err error
-	return err
-}
-
-func teardown() error {
-	var err error
-	return err
-}
