+++ szproduct/szproduct.go
@@ -29,7 +29,7 @@ import (
 	"github.com/senzing-garage/go-observing/notifier"
 	"github.com/senzing-garage/go-observing/observer"
 	"github.com/senzing-garage/go-observing/subject"
-	"github.com/senzing-garage/sz-sdk-go-core/helpers"
+	"github.com/senzing-garage/sz-sdk-go-core/helper"
 	"github.com/senzing-garage/sz-sdk-go/szerror"
 	szproductapi "github.com/senzing-garage/sz-sdk-go/szproduct"
 )
@@ -349,7 +349,7 @@ func (client *Szproduct) init(ctx context.Context, instanceName string, settings
 // Get the Logger singleton.
 func (client *Szproduct) getLogger() logging.Logging {
 	if client.logger == nil {
-		client.logger = helpers.GetLogger(ComponentID, szproductapi.IDMessages, baseCallerSkip)
+		client.logger = helper.GetLogger(ComponentID, szproductapi.IDMessages, baseCallerSkip)
 	}
 	return client.logger
 }
@@ -357,7 +357,7 @@ func (client *Szproduct) getLogger() logging.Logging {
 // Get the Messenger singleton.
 func (client *Szproduct) getMessenger() messenger.Messenger {
 	if client.messenger == nil {
-		client.messenger = helpers.GetMessenger(ComponentID, szproductapi.IDMessages, baseCallerSkip)
+		client.messenger = helper.GetMessenger(ComponentID, szproductapi.IDMessages, baseCallerSkip)
 	}
 	return client.messenger
 }
